+++ R/ip_to_hex.R
@@ -3,15 +3,15 @@
 #' Encode or decode an [`ip_address`] as a hexadecimal string.
 #'
 #' @param x
-#'  * For `ip_to_hex()`: An [`ip_address`] vector
-#'  * For `hex_to_ip()`: A character vector containing hexadecimal strings
+#'  * `ip_to_hex()`: An [`ip_address`] vector
+#'  * `hex_to_ip()`: A character vector containing hexadecimal strings
 #' @param is_ipv6 A logical vector indicating whether to construct an IPv4 or
 #'   IPv6 address. If `NULL` (the default), then IPv4 is preferred but an IPv6
 #'   address is constructed when `x` is too large for the IPv4 address space.
 #'
 #' @return
-#'  * For `ip_to_hex()`: A character vector
-#'  * For `hex_to_ip()`: An [`ip_address`] vector
+#'  * `ip_to_hex()`: A character vector
+#'  * `hex_to_ip()`: An [`ip_address`] vector
 #'
 #' @examples
 #' x <- ip_address(c("192.168.0.1", "2001:db8::8a2e:370:7334", NA))
@@ -21,25 +21,21 @@
 #' @family address representations
 #' @export
 ip_to_hex <- function(x) {
-  if (!is_ip_address(x)) {
-    abort("`x` must be an ip_address vector")
-  }
-
+  check_address(x)
   wrap_encode_hex(x)
 }
 
 #' @rdname ip_to_hex
 #' @export
 hex_to_ip <- function(x, is_ipv6 = NULL) {
-  if (!is_character(x)) {
-    abort("`x` must be a character vector")
-  }
-  if (!(is_null(is_ipv6) || is_logical(is_ipv6))) {
-    abort("`is_ipv6` must be a logical vector or NULL")
-  }
+  check_character(x)
+  check_all(
+    grepl("^0[xX][0-9a-fA-F]+$", x %|% "0xff", perl = TRUE),
+    "x", "must be a hexadecimal string"
+  )
 
-  if (!all(grepl("^0[xX][0-9a-fA-F]+$", x[!is.na(x)], perl = TRUE))) {
-    abort("Found invalid hexadecimal string")
+  if (!(is_null(is_ipv6) || is_logical(is_ipv6))) {
+    cli::cli_abort("{.arg is_ipv6} must be a logical vector or NULL")
   }
 
   if (is_null(is_ipv6)) {