+++ R/network_mask.R
@@ -10,15 +10,21 @@
 #' [`ip_address`] vector of netmasks and/or hostmasks, while `netmask()` and
 #' `hostmask()` can accept a vector of prefix lengths.
 #'
-#' @param x An [`ip_network`] or [`ip_interface`] vector
-#' @param mask An [`ip_address`] vector of netmasks and/or hostmasks. Ambiguous
-#'   cases (all zeros, all ones) are treated as netmasks.
-#' @param prefix_length An integer vector
-#' @param is_ipv6 A logical vector
-#' @param ... Arguments to be passed to other methods
+#' @param x
+#'  * An [`ip_network`] vector.
+#'  * An [`ip_interface`] vector.
+#'  * `prefix_length()`: An [`ip_address`] vector of netmasks and/or hostmasks. Ambiguous cases
+#'  (all zeros, all ones) are treated as netmasks.
+#'  * `netmask()` and `hostmask()`: An integer vector of prefix lengths.
+#' @param is_ipv6 A logical vector indicating whether to construct an IPv4 or
+#'   IPv6 address. If `NULL` (the default), then IPv4 is preferred but an IPv6
+#'   address is constructed when `x` is too large for the IPv4 address space.
+#' @inheritParams rlang::args_dots_used
+#'
 #' @return
-#' * `prefix_length()` returns an integer vector
-#' * `netmask()` and `hostmask()` return an [`ip_address`] vector
+#' * `prefix_length()`: An integer vector
+#' * `netmask()`: An [`ip_address`] vector
+#' * `hostmask()`: An [`ip_address`] vector
 #'
 #' @examples
 #' x <- ip_network(c("192.168.0.0/22", "2001:db00::0/26"))
@@ -44,33 +50,49 @@
 #' @name netmask
 NULL
 
+
+# prefix_length ----------------------------------------------------------------
+
 #' @rdname netmask
 #' @export
-prefix_length <- function(...) {
+prefix_length <- function(x) {
   UseMethod("prefix_length")
 }
 
-#' @rdname netmask
 #' @export
-netmask <- function(...) {
-  UseMethod("netmask")
-}
-
-#' @rdname netmask
-#' @export
-hostmask <- function(...) {
-  UseMethod("hostmask")
-}
-
-#' @rdname netmask
-#' @export
-prefix_length.ip_network <- function(x, ...) {
+prefix_length.ip_network <- function(x) {
   # directly returning field enables assignment
   result <- field(x, "prefix")
   result
 }
 
+#' @export
+prefix_length.ip_interface <- function(x) {
+  # directly returning field enables assignment
+  result <- field(x, "prefix")
+  result
+}
+
+#' @export
+prefix_length.ip_address <- function(x) {
+  wrap_prefix_from_mask(x)
+}
+
+#' @export
+prefix_length.default <- function(x) {
+  cli::cli_abort("{.arg x} must be an {.cls ip_network}, {.cls ip_interface} or {.cls ip_address} vector")
+}
+
+
+# netmask ----------------------------------------------------------------------
+
 #' @rdname netmask
+#' @export
+netmask <- function(x, ...) {
+  check_dots_used()
+  UseMethod("netmask")
+}
+
 #' @export
 netmask.ip_network <- function(x, ...) {
   wrap_netmask(
@@ -79,24 +101,6 @@ netmask.ip_network <- function(x, ...) {
   )
 }
 
-#' @rdname netmask
-#' @export
-hostmask.ip_network <- function(x, ...) {
-  wrap_hostmask(
-    field(x, "prefix"),
-    field(x, "is_ipv6")
-  )
-}
-
-#' @rdname netmask
-#' @export
-prefix_length.ip_interface <- function(x, ...) {
-  # directly returning field enables assignment
-  result <- field(x, "prefix")
-  result
-}
-
-#' @rdname netmask
 #' @export
 netmask.ip_interface <- function(x, ...) {
   wrap_netmask(
@@ -107,6 +111,34 @@ netmask.ip_interface <- function(x, ...) {
 
 #' @rdname netmask
 #' @export
+netmask.numeric <- function(x, is_ipv6 = NULL, ...) {
+  subnet_mask(x, is_ipv6, wrap_netmask)
+}
+
+#' @export
+netmask.default <- function(x, ...) {
+  cli::cli_abort("{.arg x} must be an {.cls ip_network}, {.cls ip_interface} or integer vector")
+}
+
+
+# hostmask ---------------------------------------------------------------------
+
+#' @rdname netmask
+#' @export
+hostmask <- function(x, ...) {
+  check_dots_used()
+  UseMethod("hostmask")
+}
+
+#' @export
+hostmask.ip_network <- function(x, ...) {
+  wrap_hostmask(
+    field(x, "prefix"),
+    field(x, "is_ipv6")
+  )
+}
+
+#' @export
 hostmask.ip_interface <- function(x, ...) {
   wrap_hostmask(
     field(x, "prefix"),
@@ -116,48 +148,37 @@ hostmask.ip_interface <- function(x, ...) {
 
 #' @rdname netmask
 #' @export
-prefix_length.default <- function(mask, ...) {
-  if (!is_ip_address(mask)) {
-    abort("prefix_length() accepts an ip_address, ip_network or ip_interface vector")
-  }
-
-  wrap_prefix_from_mask(mask)
+hostmask.numeric <- function(x, is_ipv6 = NULL, ...) {
+  subnet_mask(x, is_ipv6, wrap_hostmask)
 }
 
-#' @rdname netmask
 #' @export
-netmask.default <- function(prefix_length, is_ipv6, ...) {
-  subnet_mask(prefix_length, is_ipv6, wrap_netmask)
+hostmask.default <- function(x, ...) {
+  cli::cli_abort("{.arg x} must be an {.cls ip_network}, {.cls ip_interface} or integer vector")
 }
 
-#' @rdname netmask
-#' @export
-hostmask.default <- function(prefix_length, is_ipv6, ...) {
-  subnet_mask(prefix_length, is_ipv6, wrap_hostmask)
-}
+
+# prefix to mask ---------------------------------------------------------------
 
 subnet_mask <- function(prefix_length, is_ipv6, mask_func) {
-  if (!is_integerish(prefix_length)) {
-    abort("`prefix_length` must be an integer vector")
-  }
-  if (!is_logical(is_ipv6)) {
-    abort("`is_ipv6` must be a logical vector")
+  force(is_ipv6)
+  check_integer(prefix_length)
+
+  if (!(is_null(is_ipv6) || is_logical(is_ipv6))) {
+    cli::cli_abort("{.arg is_ipv6} must be a logical vector or NULL")
   }
 
-  # vector recycling
-  args <- vec_recycle_common(prefix_length, is_ipv6)
-  prefix_length <- args[[1L]]
-  is_ipv6 <- args[[2L]]
+  if (is_null(is_ipv6)) {
+    is_ipv6 <- prefix_length > 32L
+  } else {
+    args <- vec_recycle_common(prefix_length, is_ipv6)
+    prefix_length <- args[[1L]]
+    is_ipv6 <- args[[2L]]
+  }
 
-  if (any(prefix_length < 0L, na.rm = TRUE)) {
-    abort("`prefix_length` cannot be negative")
-  }
-  if (any(prefix_length[!is_ipv6] > 32L, na.rm = TRUE)) {
-    abort("`prefix_length` cannot be greater than 32 for IPv4")
-  }
-  if (any(prefix_length[is_ipv6] > 128L, na.rm = TRUE)) {
-    abort("`prefix_length` cannot be greater than 128 for IPv6")
-  }
+  check_all(prefix_length >=   0L,            "prefix_length", "must be positive")
+  check_all(prefix_length <=  32L |  is_ipv6, "prefix_length", "must not be greater than 32 for IPv4")
+  check_all(prefix_length <= 128L | !is_ipv6, "prefix_length", "must not be greater than 128 for IPv6")
 
   do.call(mask_func, list(prefix_length, is_ipv6))
 }