+++ R/network_size.R
@@ -6,9 +6,9 @@
 #'
 #' @param x An [`ip_network`] vector
 #' @return
-#' * `network_address()` and `broadcast_address()` return an
-#'   [`ip_address`] vector
-#' * `num_addresses()` returns a numeric vector
+#' * `network_address()`: An [`ip_address`] vector
+#' * `broadcast_address()`: An [`ip_address`] vector
+#' * `num_addresses()`: A numeric vector
 #'
 #' @examples
 #' x <- ip_network(c("192.168.0.0/22", "2001:db8::/33"))
@@ -26,10 +26,7 @@ NULL
 #' @rdname network_size
 #' @export
 network_address <- function(x) {
-  if (!is_ip_network(x)) {
-    abort("`x` must be an ip_network vector")
-  }
-
+  check_network(x)
   new_ip_address(
     field(x, "address1"),
     field(x, "address2"),
@@ -50,19 +47,13 @@ network_address <- function(x) {
 #' @rdname network_size
 #' @export
 broadcast_address <- function(x) {
-  if (!is_ip_network(x)) {
-    abort("`x` must be an ip_network vector")
-  }
-
+  check_network(x)
   wrap_broadcast_address(x)
 }
 
 #' @rdname network_size
 #' @export
 num_addresses <- function(x) {
-  if (!is_ip_network(x)) {
-    abort("`x` must be an ip_network vector")
-  }
-
+  check_network(x)
   2L^(max_prefix_length(x) - prefix_length(x))
 }