+++ R/sequence.R
@@ -1,7 +1,7 @@
 #' List addresses within a network
 #'
 #' @param x An [`ip_network`] scalar
-#' @param ... Included for generic consistency
+#' @inheritParams rlang::args_dots_empty
 #' @return An [`ip_address`] vector
 #'
 #' @details
@@ -43,9 +43,8 @@ NULL
 #' @rdname sequence
 #' @export
 seq.ip_network <- function(x, ...) {
-  if (length(x) != 1) {
-    abort("`x` must be an ip_network scalar")
-  }
+  check_dots_empty()
+  check_scalar(x)
   if (any(max_prefix_length(x) - prefix_length(x) > 30L, na.rm = TRUE)) {
     abort("Network too large")
   }
@@ -59,9 +58,8 @@ seq.ip_network <- function(x, ...) {
 #' @rdname sequence
 #' @export
 hosts <- function(x) {
-  if (!(is_ip_network(x) && length(x) == 1)) {
-    abort("`x` must be an ip_network scalar")
-  }
+  check_network(x)
+  check_scalar(x)
   if (any(max_prefix_length(x) - prefix_length(x) > 30L, na.rm = TRUE)) {
     abort("Network too large")
   }