+++ NEWS.md
@@ -1,18 +1,47 @@
+# ipaddress 1.0.0
+
+To mark the release of ipaddress 1.0.0, some **breaking changes** were made to improve the interface for data analysis.
+
+## Breaking changes
+
+* Removed the `multiple` argument from `ip_to_hostname()` and `hostname_to_ip()` (#83).
+  * `ip_to_hostname()` and `hostname_to_ip()` now always return one result per input (stored in a vector).
+  * New `ip_to_hostname_all()` and `hostname_to_ip_all()` return all results per input (stored in a list of vectors).
+* `subnets()` now takes vector input and outputs a list of vectors (#83).
+* Some arguments must now be explicitly passed using the argument name:
+  * The `strict` argument of `ip_network()`.
+  * The `exploded` argument of `format.ip_address()`, `format.ip_network()` and `format.ip_interface()`.
+  * The `replace` argument of `sample_ipv4()`, `sample_ipv6()` and `sample_network()`.
+
+## New features
+
+* New `country_networks()` downloads IP networks registered to specific countries (#87).
+* Error formatting has been improved using {cli} (#86).
+
+## Minor improvements and fixes
+
+* The `is_ipv6` argument of `netmask()` and `hostmask()` defaults to `NULL`. This option prefers an IPv4 mask, but creates an IPv6 mask as a fallback.
+* Renamed `vignette("ipaddress-classes")` to `vignette("ip-data")`.
+* Renamed `vignette("ipaddress-examples")` to `vignette("recipes")`.
+* Updated the `iana_ipv4` dataset.
+* Tests now use {testthat} 3e.
+
+
 # ipaddress 0.5.6
 
-Resolve CRAN check warnings about deprecated use of C++ function `sprintf()`.
+Fix for CRAN checks.
 
 # ipaddress 0.5.5
 
-Resolve CRAN check notes about HTML5 by upgrading to latest {roxygen}.
+Fix for CRAN checks.
 
 # ipaddress 0.5.4
 
-Compatible with new compiler toolchain on Windows for R 4.2 ([see announcement](https://developer.r-project.org/Blog/public/2021/12/07/upcoming-changes-in-r-4.2-on-windows/)).
+Fix for CRAN checks.
 
 # ipaddress 0.5.3
 
-Hotfix for CRAN check warnings.
+Fix for CRAN checks.
 
 
 # ipaddress 0.5.2
@@ -170,7 +199,7 @@ This release achieves feature parity with the Python [ipaddress](https://docs.py
 
 # ipaddress 0.1.1
 
-Hotfix for CRAN errors on Solaris.
+Fix for CRAN checks.
 
 
 # ipaddress 0.1.0